Bugfix: Title input width not constrained (#10)
* Adding a maxLength on the title input field

* New better max width approach

The new approach alternates the display for a text label and the input field text. This enables us to support any length text string. By default the label truncates. But when the input field is focused, you can scroll within the field to edit the whole thing.

* Ignore the redundant label for screen readers

label is the same as the value of the input, so it doesn't need to be read twice.